requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx>=0.27.0
redis>=5.0.0
celery>=5.3.0
google-genai>=1.0.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.126.0
supabase>=2.7.0
boto3>=1.34.0
python-multipart==0.0.9
structlog>=24.1.0
python-dotenv>=1.0.0
Pillow>=10.3.0
jinja2>=3.1.4
aiofiles>=23.2.1
tenacity>=8.2.3
